requirement.txt
absl-py==2.3.1
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
flatbuffers==20181003210633
fonttools==4.59.0
jax==0.7.0
jaxlib==0.7.0
kiwisolver==1.4.8
matplotlib==3.10.3
mediapipe==0.10.18
ml_dtypes==0.5.1
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
RPi.GPIO==0.7.1
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
